View

+++ app/src/main/java/com/example/weatherassignment/database/model/ResultDateAndValue.kt
@@ -1,3 +1,3 @@
 package com.example.weatherassignment.database.model
 
-data class ResultDateAndValue(val year:Int, val month:Int, val day:Int, val value:Float )
+data class ResultDateAndValue(val year: Int, val month: Int, val day: Int, val value: Int)
